refactor: chain detailed
