+++ packages/maskbook/src/plugins/Wallet/services/assets.ts
@@ -8,9 +8,9 @@ import {
     CurrencyType,
     EthereumTokenType,
     formatEthereumAddress,
-    getChainDetailed,
     getChainIdFromName,
     getTokenConstants,
+    isChainMainnet,
     NetworkType,
     pow10,
 } from '@masknet/web3-shared'
@@ -101,32 +101,28 @@ export async function getAssetsList(
 }
 
 function formatAssetsFromDebank(data: BalanceRecord[]) {
-    return data.map((x): Asset => {
-        const chainId = getChainIdFromName(x.id)
-        return {
-            chain: x.chain,
-            token:
-                chainId && getChainDetailed(chainId)?.network === 'mainnet'
-                    ? createNativeToken(getChainIdFromName(x.id) ?? ChainId.Mainnet)
-                    : createERC20Token(
-                          getChainIdFromName(x.chain) ?? ChainId.Mainnet,
-                          formatEthereumAddress(x.id),
-                          x.decimals,
-                          x.name,
-                          x.symbol,
-                      ),
-            balance: new BigNumber(x.balance).toFixed(),
-            price: {
-                [CurrencyType.USD]: new BigNumber(x.price).toFixed(),
-            },
-            value: {
-                [CurrencyType.USD]: new BigNumber(x.price)
-                    .multipliedBy(new BigNumber(x.balance).dividedBy(pow10(x.decimals)))
-                    .toFixed(),
-            },
-            logoURL: x.logo_url,
-        }
-    })
+    return data
+        .filter((x) => getChainIdFromName(x.id))
+        .map((y): Asset => {
+            const chainId = getChainIdFromName(y.id) ?? ChainId.Mainnet
+            const chainId_ = getChainIdFromName(y.chain) ?? ChainId.Mainnet
+            return {
+                chain: y.chain,
+                token: isChainMainnet(chainId)
+                    ? createNativeToken(chainId)
+                    : createERC20Token(chainId_, formatEthereumAddress(y.id), y.decimals, y.name, y.symbol),
+                balance: new BigNumber(y.balance).toFixed(),
+                price: {
+                    [CurrencyType.USD]: new BigNumber(y.price).toFixed(),
+                },
+                value: {
+                    [CurrencyType.USD]: new BigNumber(y.price)
+                        .multipliedBy(new BigNumber(y.balance).dividedBy(pow10(y.decimals)))
+                        .toFixed(),
+                },
+                logoURL: y.logo_url,
+            }
+        })
 }
 
 const filterAssetType = ['compound', 'trash', 'uniswap', 'uniswap-v2', 'nft']
