+++ packages/maskbook/src/web3/UI/EthereumChainBoundary.tsx
@@ -7,11 +7,9 @@ import {
     getNetworkTypeFromChainId,
     NetworkType,
     ProviderType,
-    resolveProviderName,
     useAccount,
-    useAllowTestnet,
-    useChainDetailed,
     useChainId,
+    useChainIdValid,
 } from '@masknet/web3-shared'
 import { useValueRef, delay } from '@masknet/shared'
 import { ActionButtonPromise } from '../../extension/options-page/DashboardComponents/ActionButton'
@@ -29,8 +27,7 @@ export function EthereumChainBoundary(props: EthereumChainBoundaryProps) {
     const { t } = useI18N()
     const account = useAccount()
     const chainId = useChainId()
-    const chainDetailed = useChainDetailed()
-    const allowTestnet = useAllowTestnet()
+    const chainIdValid = useChainIdValid()
     const providerType = useValueRef(currentProviderSettings)
 
     const expectedChainId = props.chainId
@@ -38,11 +35,8 @@ export function EthereumChainBoundary(props: EthereumChainBoundaryProps) {
     const acutalChainId = chainId
     const actualNetwork = getChainName(acutalChainId)
 
-    // if false then the user should switch network manually
-    const isSwitchable = true
-
-    // if testnets were not allowed it will not guide the user to switch the network
-    const isAllowed = allowTestnet || chainDetailed?.network === 'mainnet'
+    // if false then it will not guide the user to switch the network
+    const isAllowed = chainIdValid && !!account
 
     const onSwitch = useCallback(async () => {
         // a short time loading makes the user fells better
@@ -67,8 +61,9 @@ export function EthereumChainBoundary(props: EthereumChainBoundaryProps) {
         if (!networkType) return
         if (networkType === NetworkType.Ethereum) await Services.Ethereum.switchEthereumChain(expectedChainId)
         else await Services.Ethereum.addEthereumChain(chainDetailedCAIP, account)
-    }, [account, isAllowed, isSwitchable, providerType, expectedChainId])
+    }, [account, isAllowed, providerType, expectedChainId])
 
+    // matched
     if (acutalChainId === expectedChainId) return <>{props.children}</>
 
     if (!isAllowed)
@@ -92,16 +87,8 @@ export function EthereumChainBoundary(props: EthereumChainBoundaryProps) {
                         network: actualNetwork,
                     })}
                 </span>
-                {isSwitchable ? null : (
-                    <span>
-                        {t('plugin_wallet_swtich_to', {
-                            network: expectedNetwork,
-                            provider: resolveProviderName(providerType),
-                        })}
-                    </span>
-                )}
             </Typography>
-            {isSwitchable ? (
+            {isAllowed ? (
                 <ActionButtonPromise
                     variant="contained"
                     size="small"
