+++ packages/web3-shared/src/hooks/useChainIdValid.ts
@@ -1,0 +1,5 @@
+import { useWeb3StateContext } from '../context'
+
+export function useChainIdValid() {
+    return useWeb3StateContext().chainIdValid
+}
