+++ packages/web3-shared/src/utils/chainDetailed.ts
@@ -1,0 +1,92 @@
+import { safeUnreachable } from '@dimensiondev/kit'
+import CHAINS from '../assets/chains.json'
+import { getRPCConstants } from '../constants'
+import { ChainId, NetworkType } from '../types'
+
+export function isChainIdValid(chainId: ChainId, allowTestnet = false) {
+    const chainDetailed = getChainDetailed(chainId)
+    return !!getNetworkTypeFromChainId(chainId) && (chainDetailed?.network === 'mainnet' || allowTestnet)
+}
+
+export function isChainMainnet(chainId: ChainId) {
+    const chainDetailed = getChainDetailed(chainId)
+    return chainDetailed?.network === 'mainnet'
+}
+
+export function getChainDetailed(chainId = ChainId.Mainnet) {
+    return CHAINS.find((x) => x.chainId === chainId)
+}
+
+// Learn more: https://github.com/ChainAgnostic/CAIPs/blob/master/CAIPs/caip-2.md
+export function getChainDetailedCAIP(chainId = ChainId.Mainnet) {
+    const chainDetailed = getChainDetailed(chainId)
+    const { RPC } = getRPCConstants(chainId)
+    if (!chainDetailed) return
+    return {
+        chainId: `0x${chainDetailed.chainId.toString(16)}`,
+        chainName: chainDetailed.name,
+        nativeCurrency: chainDetailed.nativeCurrency,
+        rpcUrls: RPC,
+        blockExplorerUrls: [
+            chainDetailed.explorers && chainDetailed.explorers.length > 0 && chainDetailed.explorers[0].url
+                ? chainDetailed.explorers[0].url
+                : chainDetailed.infoURL,
+        ],
+    }
+}
+
+export function getChainRPC(chainId: ChainId, seed: number) {
+    const { RPC, RPC_WEIGHTS } = getRPCConstants(chainId)
+    if (!RPC || !RPC_WEIGHTS) throw new Error('Unknown chain id.')
+    return RPC[RPC_WEIGHTS[seed]]
+}
+
+export function getChainName(chainId: ChainId) {
+    const chainDetailed = getChainDetailed(chainId)
+    return chainDetailed?.name ?? 'Unknown Network'
+}
+
+export function getChainFullName(chainId: ChainId) {
+    const chainDetailed = getChainDetailed(chainId)
+    return chainDetailed?.fullName ?? 'Unknown Network'
+}
+
+export function getChainIdFromName(name: string) {
+    const chainDetailed = CHAINS.find((x) =>
+        [x.chain, x.network, x.name, x.shortName, x.fullName ?? '']
+            .filter(Boolean)
+            .map((y) => y.toLowerCase())
+            .includes(name.toLowerCase()),
+    )
+    return chainDetailed && getNetworkTypeFromChainId(chainDetailed.chainId)
+        ? (chainDetailed.chainId as ChainId)
+        : undefined
+}
+
+export function getChainIdFromNetworkType(networkType: NetworkType) {
+    switch (networkType) {
+        case NetworkType.Ethereum:
+            return ChainId.Mainnet
+        case NetworkType.Binance:
+            return ChainId.BSC
+        case NetworkType.Polygon:
+            return ChainId.Matic
+        default:
+            safeUnreachable(networkType)
+            return ChainId.Mainnet
+    }
+}
+
+export function getNetworkTypeFromChainId(chainId: ChainId) {
+    const chainDetailed = getChainDetailed(chainId)
+    switch (chainDetailed?.chain) {
+        case 'ETH':
+            return NetworkType.Ethereum
+        case 'BSC':
+            return NetworkType.Binance
+        case 'Matic':
+            return NetworkType.Polygon
+        default:
+            return
+    }
+}
