+++ packages/web3-shared/src/utils/index.ts
@@ -1,3 +1,5 @@
+export * from './address'
 export * from './token'
 export * from './formatter'
 export * from './number'
+export * from './chainDetailed'
